refactor: moved validator logic to the corresponding module.go file of the given route
* Moved validator logic to the corresponding module.go file of the given route
* Added Swagger documentation
* Upgraded dependencies

+++ internal/router/api/v1/note/version/model.go
@@ -6,24 +6,24 @@ import (
 
 // CreateNoteVersionRequest is the request DTO to create a note version
 type CreateNoteVersionRequest struct {
-	NoteID           uint   `json:"note_id"`
+	NoteID           int64  `json:"note_id"`
 	EncryptedContent string `json:"encrypted_content"`
 }
 
 // UpdateNoteVersionRequest is the request DTO to update a note version
 type UpdateNoteVersionRequest struct {
-	NoteVersionID    uint    `json:"note_version_id"`
+	NoteVersionID    int64   `json:"note_version_id"`
 	EncryptedContent *string `json:"encrypted_content,omitempty"`
 }
 
 // DeleteNoteVersionRequest is the request DTO to delete a note version
 type DeleteNoteVersionRequest struct {
-	NoteVersionID uint `json:"note_version_id"`
+	NoteVersionID int64 `json:"note_version_id"`
 }
 
 // GetNoteVersionRequest is the request DTO to get a note version
 type GetNoteVersionRequest struct {
-	NoteVersionID uint `json:"note_version_id"`
+	NoteVersionID int64 `json:"note_version_id"`
 }
 
 // GetNoteVersionResponse is the response DTO to get a note version
